Update default library state to material design
New look and feel of the library that follows material design
guidelines. This update of the library introduce:

- Change the default parameters of the tabs layout to make
it look like more material.

- The indicator is center by default while swiping tabs.

- The alpha value of the titles change depending the
selected position.

- Ability to change the divider width (No material change).

- Ability to pass customTabs implementing the interface
'CustomTabProvider' in your adapter (dev can use ripple
effect on the tabs).In case the customTab have
a view with id '@+id/tab_title', it will be use to place
the title and alpha will act as usual. If this interface is
not implemented then, the default Tab layout will be use (TextView).